Use float for Command.ts and update tests
Change Command.ts type from int to float to allow fractional timestamps and update the test expectation accordingly. Adjusted tests to assert a float value for ts. Also updated README device status for C1 to indicate testing is in progress.

+++ custom_components/wybot/wybot_models.py
@@ -30,7 +30,7 @@ class Command(BaseModel):
     # 9 - Data Request
     cmd: int
     dp: list[DP]
-    ts: int
+    ts: float
 
     def get_dps_as_keyed_dict(self) -> dict[str, GenericDP]:
         """Return the DP list as a keyed dictionary."""
